requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
boto3==1.34.143
Pillow==10.4.0
